Moving libmr to libmrTorch to avoid confusion with original libmr
